Rename StoredFile and StoredFiles to StorageFile and StorageFiles
